fix js files modules importing and using
