+++ templates/profile/create-post.html
@@ -52,7 +52,6 @@
                         <div class="submit-row mt-4">
                             <button id="submit-create-post" value="Зберегти" class="button-submit default primary-btn" name="_save">Зберегти</button>
                         </div>
-                        <script src="{% static 'js/create_post.js' %}" defer></script>
                         <script id="django-admin-form-add-constants" src="/static/admin/js/change_form.js" data-model-name="post" async></script>
                         <script id="django-admin-prepopulated-fields-constants" src="/static/admin/js/prepopulate_init.js" data-prepopulated-fields="[]"></script>
                     </form>
